Rename entity.py to views.py
